requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==24.3.0
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==4.2.4
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
contourpy==1.1.1
cycler==0.12.1
fonttools==4.55.4
future==1.0.0
gast==0.3.3
google-auth==2.37.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.17
grpcio==1.68.1
h5py==2.10.0
idna @ file:///croot/idna_1714398848350/work
imageio==2.9.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
lmdb==1.5.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
numpy==1.23.3
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
pandas==2.0.3
Pillow==8.2.0
protobuf==3.20.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pyswarm==0.6
pyswarms==1.3.0
python-dateutil==2.9.0.post0
pytorch-gan-metrics==0.5.1
pytz==2024.2
PyYAML==6.0.2
requests @ file:///croot/requests_1721410876868/work
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.4.1
six==1.17.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==2.4.0
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.61.1
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.2
urllib3 @ file:///croot/urllib3_1727769808118/work
Werkzeug==3.0.6
wrapt==1.17.0
zipp==3.20.2
